+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 ## Unreleased
 
+## v0.7.0-B190633 (pre-release)
+
 - Fix circular rule dependency issue. [#190](https://github.com/BernieWhite/PSRule/issues/190)
 - Fix rule `DependsOn` parameter allows null. [#191](https://github.com/BernieWhite/PSRule/issues/191)
 - Fix error message when attempting to use the rule keyword in a rule definition. [#189](https://github.com/BernieWhite/PSRule/issues/189)